server/requirements.txt
fastapi==0.95.0
PyJWT==2.6.0
python-decouple==3.8
python-multipart==0.0.6
requests==2.28.2
pydantic==1.10.6
uvicorn==0.21.1
Pillow==9.4.0
cryptography==39.0.2
pymongo==4.3.3
pyotp==2.8.0
pytz==2022.7.1
pyhanko==0.18.0
pyhanko-certvalidator==0.22.0